CropRecommendationSystem/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
requests
python-dotenv
joblib
seaborn
matplotlib
pydantic
